wip CI fixes
Former-commit-id: 93cb846dbf46d82c483508570f6014aa053ab78c

+++ src/test/compress_tests.cpp
@@ -47,10 +47,10 @@ BOOST_AUTO_TEST_CASE(compress_amounts)
     BOOST_CHECK(TestPair(21000000*COIN, 0x1406f40));
     // SYSCOIN
     auto compressed = CompressAmount(888000000*COIN);
-    BOOST_CHECK_EQUAL(888000000*COIN, DecompressAmount(compressed));
+    BOOST_CHECK_EQUAL(888000000*COIN, (int64_t)DecompressAmount(compressed));
     // max amount that can be compressed without unsigned integer overflow its a bit above 2 quintillion range
     compressed = CompressAmount((2^64)/9);
-    BOOST_CHECK_EQUAL((2^64)/9, DecompressAmount(compressed));
+    BOOST_CHECK_EQUAL((2^64)/9, (int64_t)DecompressAmount(compressed));
     for (uint64_t i = 1; i <= NUM_MULTIPLES_UNIT; i++)
         BOOST_CHECK(TestEncode(i));
